+++ test/_google_cast_context/android_google_cast_context_method_channel_test.dart
@@ -184,6 +184,7 @@ void main() {
             'suspendSessionsWhenBackgrounded': false,
             'stopReceiverApplicationWhenEndingSession': true,
             'startDiscoveryAfterFirstTapOnCastButton': false,
+            'stopCastingOnAppTerminated': false,
           }));
     });
   });